Disable by default gauge-based histograms in the Micrometer bridge
