+++ instrumentation/micrometer/micrometer-1.5/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/micrometer/v1_5/MicrometerSingletons.java
@@ -23,6 +23,9 @@ public final class MicrometerSingletons {
             .setBaseTimeUnit(
                 TimeUnitParser.parseConfigValue(
                     config.getString("otel.instrumentation.micrometer.base-time-unit")))
+            .setMicrometerHistogramGaugesEnabled(
+                config.getBoolean(
+                    "otel.instrumentation.micrometer.histogram-gauges.enabled", false))
             .build();
   }
