+++ instrumentation/micrometer/micrometer-1.5/library/src/main/java/io/opentelemetry/instrumentation/micrometer/v1_5/OpenTelemetryMeterRegistryBuilder.java
@@ -7,7 +7,10 @@ package io.opentelemetry.instrumentation.micrometer.v1_5;
 
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import io.micrometer.core.instrument.Clock;
+import io.micrometer.core.instrument.DistributionSummary;
+import io.micrometer.core.instrument.LongTaskTimer;
 import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.Timer;
 import io.micrometer.core.instrument.config.NamingConvention;
 import io.opentelemetry.api.OpenTelemetry;
 import java.util.concurrent.TimeUnit;
@@ -22,6 +25,7 @@ public final class OpenTelemetryMeterRegistryBuilder {
   private Clock clock = Clock.SYSTEM;
   private TimeUnit baseTimeUnit = TimeUnit.MILLISECONDS;
   private boolean prometheusMode = false;
+  private boolean histogramGaugesEnabled = false;
 
   OpenTelemetryMeterRegistryBuilder(OpenTelemetry openTelemetry) {
     this.openTelemetry = openTelemetry;
@@ -55,6 +59,27 @@ public final class OpenTelemetryMeterRegistryBuilder {
   }
 
   /**
+   * Enables the generation of gauge-based Micrometer histograms. While the Micrometer bridge is
+   * able to map Micrometer's {@link DistributionSummary} and {@link Timer} service level objectives
+   * to OpenTelemetry histogram buckets, it might not cover all cases that are normally supported by
+   * Micrometer (e.g. the bridge is not able to translate percentiles). With this setting enabled,
+   * the Micrometer bridge will additionally emit Micrometer service level objectives and
+   * percentiles as separate gauges.
+   *
+   * <p>Note that this setting does not concern the {@link LongTaskTimer}, as it is not bridged to
+   * an OpenTelemetry histogram.
+   *
+   * <p>This is disabled by default, set this to {@code true} to enable gauge-based Micrometer
+   * histograms.
+   */
+  @CanIgnoreReturnValue
+  public OpenTelemetryMeterRegistryBuilder setMicrometerHistogramGaugesEnabled(
+      boolean histogramGaugesEnabled) {
+    this.histogramGaugesEnabled = histogramGaugesEnabled;
+    return this;
+  }
+
+  /**
    * Returns a new {@link OpenTelemetryMeterRegistry} with the settings of this {@link
    * OpenTelemetryMeterRegistryBuilder}.
    */
@@ -63,11 +88,16 @@ public final class OpenTelemetryMeterRegistryBuilder {
     TimeUnit baseTimeUnit = prometheusMode ? TimeUnit.SECONDS : this.baseTimeUnit;
     NamingConvention namingConvention =
         prometheusMode ? PrometheusModeNamingConvention.INSTANCE : NamingConvention.identity;
+    DistributionStatisticConfigModifier modifier =
+        histogramGaugesEnabled
+            ? DistributionStatisticConfigModifier.IDENTITY
+            : DistributionStatisticConfigModifier.DISABLE_HISTOGRAM_GAUGES;
 
     return new OpenTelemetryMeterRegistry(
         clock,
         baseTimeUnit,
         namingConvention,
+        modifier,
         openTelemetry.getMeterProvider().get(INSTRUMENTATION_NAME));
   }
 }
